reorganize check kernel output tests
